refactor: 패키지 구조 변경 - 원활한 화면전환을 위해 패키지 구조 변경 - 불필요한 파일 삭제
