+++ app/src/main/java/com/openstudy/carefull/chatbot/ChatbotScreen.kt
@@ -25,7 +25,6 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.openstudy.carefull.screen.BasalMetabolicRateMeasurement
 import com.openstudy.carefull.ui.theme.CarefullTheme
 
 @Composable
