--- app/src/main/java/com/openstudy/carefull/screen/FoodInformation.kt
@@ -1,158 +1,0 @@
-package com.openstudy.carefull.screen
-
-import android.annotation.SuppressLint
-import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxHeight
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.automirrored.filled.KeyboardArrowLeft
-import androidx.compose.material3.Button
-import androidx.compose.material3.ButtonDefaults
-import androidx.compose.material3.HorizontalDivider
-import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
-import com.openstudy.carefull.ui.theme.CarefullTheme
-
-@SuppressLint("UnusedMaterial3ScaffoldPaddingParameter")
-@Preview(showBackground = true)
-@Composable
-fun FoodInformation() {
-    CarefullTheme {
-        Scaffold(
-            topBar = {
-                Column() {
-                    Row(
-                        modifier = Modifier.fillMaxWidth(),
-                        verticalAlignment = Alignment.CenterVertically
-                    ) {
-                        IconButton(onClick = {}
-                        ) {
-                            Icon(
-                                imageVector = Icons.AutoMirrored.Filled.KeyboardArrowLeft,
-                                contentDescription = null,
-                                modifier = Modifier
-                                    .size(48.dp)
-                                    .padding(top = 5.dp),
-                                tint = MaterialTheme.colorScheme.primary
-                            )
-                        }
-                        Text(
-                            text = "음식 정보",
-                            style = MaterialTheme.typography.titleLarge,
-                            modifier = Modifier.weight(1f),
-                            textAlign = TextAlign.Center
-                        )
-                        Spacer(modifier = Modifier.width(48.dp))
-                    }
-                    Spacer(modifier = Modifier.height(10.dp))
-                    HorizontalDivider(
-                        modifier = Modifier,
-                        thickness = 3.dp,
-                        color = MaterialTheme.colorScheme.primary
-                    )
-                }
-            },
-            bottomBar = {
-                Column(
-                    modifier = Modifier.fillMaxWidth(),
-                    horizontalAlignment = Alignment.CenterHorizontally
-                )
-                {
-                    HorizontalDivider(
-                        modifier = Modifier,
-                        thickness = 3.dp,
-                        color = MaterialTheme.colorScheme.primary
-                    )
-                    Spacer(modifier = Modifier.height(10.dp))
-                    Button(
-                        onClick = {
-                        },
-                        modifier = Modifier
-                            .fillMaxWidth(0.7f)
-                            .height(56.dp),
-
-                        shape = RoundedCornerShape(12.dp),
-                        colors = ButtonDefaults.buttonColors(
-                            contentColor = Color.White
-                        ),
-                    ) {
-                        Text(
-                            text = "등록하기",
-                            style = MaterialTheme.typography.labelLarge
-                        )
-                    }
-                    Spacer(modifier = Modifier.height(10.dp))
-                }
-            }
-        ) { innerPadding ->
-            Column(
-                modifier = Modifier
-                    .padding(innerPadding)
-                    .fillMaxSize(),
-                horizontalAlignment = Alignment.CenterHorizontally
-            ) {
-                Spacer(modifier = Modifier.height(20.dp))
-                Box(
-                    modifier = Modifier
-                        .fillMaxWidth(0.8f)
-                        .fillMaxHeight(0.4f)
-                        .background(color = Color.Gray)
-                )
-                Spacer(modifier = Modifier.height(20.dp))
-                HorizontalDivider(
-                    modifier = Modifier,
-                    thickness = 3.dp,
-                    color = MaterialTheme.colorScheme.primary
-                )
-                LazyColumn(
-                    modifier = Modifier
-                        .fillMaxSize()
-                        .padding(),
-                    horizontalAlignment = Alignment.CenterHorizontally,
-                ) {
-
-                    //더미데이터
-                    item {
-                        FoodSearchResultItem(
-                            food = Food(101, "햄버거", 550, 45, 25, 30),
-                            onAddClick = {})
-                        FoodSearchResultItem(
-                            food = Food(103, "치즈버거", 650, 48, 28, 35),
-                            onAddClick = {})
-                        FoodSearchResultItem(
-                            food = Food(104, "더블버거", 850, 50, 40, 50),
-                            onAddClick = {})
-                        FoodSearchResultItem(
-                            food = Food(105, "새우버거", 480, 55, 18, 22),
-                            onAddClick = {})
-                        FoodSearchResultItem(
-                            food = Food(106, "치킨버거", 520, 47, 26, 25),
-                            onAddClick = {})
-                    }
-                }
-            }
-        }
-    }
-}
