+++ app/src/main/java/com/openstudy/carefull/screen/auth/Signin.kt
@@ -1,4 +1,4 @@
-package com.openstudy.carefull.screen
+package com.openstudy.carefull.screen.auth
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.Arrangement
@@ -31,11 +31,9 @@ import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.input.KeyboardType
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.openstudy.carefull.R
 import com.openstudy.carefull.navigation.AuthScreen
-import com.openstudy.carefull.ui.theme.CarefullTheme
 
 @Composable
 fun Signin(select: (AuthScreen) -> Unit) {
