+++ app/src/main/java/com/openstudy/carefull/screen/mypage/AccountManagement.kt
@@ -1,11 +1,14 @@
-package com.openstudy.carefull.screen
+package com.openstudy.carefull.screen.mypage
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.border
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
@@ -13,6 +16,7 @@ import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
+import androidx.compose.material3.TextButton
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
@@ -23,16 +27,14 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
 import com.openstudy.carefull.R
 import com.openstudy.carefull.common.BottomNavigationBar
-import com.openstudy.carefull.ui.theme.CarefullTheme
 import com.openstudy.carefull.common.MenuButton
 import com.openstudy.carefull.common.RowLine
-import com.openstudy.carefull.common.SwitchMenuButton
+import com.openstudy.carefull.ui.theme.CarefullTheme
 
 @Composable
-fun MyPage() {
+fun AccountManagement() {
     Scaffold(
         bottomBar = {
             BottomNavigationBar(currentRoute = R.string.mypage)
@@ -44,13 +46,13 @@ fun MyPage() {
                 .padding(innerPadding)
                 .padding(top = 32.dp),
             horizontalAlignment = Alignment.CenterHorizontally,
-            verticalArrangement = Arrangement.spacedBy(10.dp)
+            verticalArrangement = Arrangement.spacedBy(20.dp)
         ) {
-
             Text(
-                text = stringResource(R.string.mypage),
+                text = stringResource(R.string.account_management),
                 style = MaterialTheme.typography.titleLarge
             )
+
             //임시 프로필 아이콘
             Image(
                 painter = painterResource(id = R.drawable.app_logo),
@@ -60,26 +62,50 @@ fun MyPage() {
                     .clip(CircleShape)
                     .border(1.dp, Color.Gray, CircleShape)
             )
-            Spacer(modifier = Modifier.height(16.dp))
+
+            Text(
+                text = stringResource(id = R.string.nickname),
+                fontWeight = FontWeight.Bold,
+                color = Color.Black
+            )
+            Text(
+                text = stringResource(id = R.string.email),
+                color = Color.Black
+            )
             RowLine()
-            MenuButton(text = stringResource(R.string.account_management), onClick = { })
+            MenuButton(text = stringResource(R.string.change_profile), onClick = { })
             RowLine()
-            MenuButton(
-                text = stringResource(R.string.basal_metabolic_rate_measurement),
-                onClick = { })
+            MenuButton(text = stringResource(R.string.change_nickname), onClick = { })
             RowLine()
-            SwitchMenuButton(text = stringResource(R.string.notification_setting))
+            MenuButton(text = stringResource(R.string.change_password), onClick = { })
             RowLine()
-            MenuButton(text = stringResource(R.string.writing_management), onClick = { })
-            RowLine()
+            Spacer(modifier = Modifier.height(80.dp))
+
+            Row(
+                modifier = Modifier
+                    .fillMaxWidth(),
+                verticalAlignment = Alignment.CenterVertically
+            ) {
+                Spacer(modifier = Modifier.fillMaxWidth(0.7f))
+                TextButton(
+                    onClick = {},
+                    contentPadding = PaddingValues(horizontal = 4.dp, vertical = 4.dp),
+                ) {
+                    Text(
+                        text = stringResource(R.string.delete_account),
+                        style = MaterialTheme.typography.labelLarge,
+                        color = MaterialTheme.colorScheme.outline
+                    )
+                }
+            }
         }
     }
 }
 
 @Preview(showBackground = true)
 @Composable
-fun MyPageScreenPreview() {
+fun AccountManagementPreview() {
     CarefullTheme {
-        MyPage()
+        AccountManagement()
     }
 }
