+++ app/src/main/java/com/openstudy/carefull/screen/routine/Diet.kt
@@ -1,0 +1,163 @@
+package com.openstudy.carefull.screen.routine
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.AccountBox
+import androidx.compose.material.icons.filled.AddCircle
+import androidx.compose.material3.Card
+import androidx.compose.material3.CardDefaults
+import androidx.compose.material3.HorizontalDivider
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import com.openstudy.carefull.ui.theme.CarefullTheme
+
+
+
+@Composable
+fun Diet() {
+    var addedFoods by remember { mutableStateOf<List<AddedFood>>(emptyList()) }
+
+    Column {
+        Spacer(modifier = Modifier.height(24.dp))
+        Text(
+            text = "오늘 식사 내역", style = MaterialTheme.typography.bodyLarge,
+            modifier = Modifier
+                .fillMaxWidth()
+                .padding(horizontal = 16.dp),
+            textAlign = TextAlign.Center
+        )
+        LazyColumn(
+            modifier = Modifier
+                .fillMaxSize()
+                .padding(16.dp)
+        ) {
+            item {
+                MealType.entries.forEach { mealType ->
+                    MealSection(
+                        mealType = mealType,
+                        addedFoods = addedFoods.filter { it.mealType == mealType },
+                        onAddClick = {
+                        },
+                        onRemoveClick = { foodToRemove ->
+                        }
+                    )
+                }
+            }
+        }
+    }
+}
+
+@Composable
+fun MealSection(
+    mealType: MealType,
+    addedFoods: List<AddedFood>,
+    onAddClick: () -> Unit,
+    onRemoveClick: (AddedFood) -> Unit
+) {
+    Card(
+        modifier = Modifier
+            .fillMaxWidth()
+            .padding(vertical = 12.dp),
+        elevation = CardDefaults.cardElevation(defaultElevation = 2.dp)
+    ) {
+        Column(
+            modifier = Modifier
+                .fillMaxWidth()
+        ) {
+            Row(
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .padding(8.dp),
+                verticalAlignment = Alignment.CenterVertically
+            ) {
+                Text(
+                    text = mealType.time,
+                    fontSize = 16.sp,
+                    fontWeight = FontWeight.Bold
+                )
+                Spacer(modifier = Modifier.weight(1f))
+                IconButton(onClick = onAddClick) {
+                    Icon(
+                        imageVector = Icons.Default.AccountBox,
+                        contentDescription = "${mealType.time} 음식 추가",
+                        tint = MaterialTheme.colorScheme.primary,
+                        modifier = Modifier.size(28.dp)
+                    )
+                }
+                IconButton(onClick = onAddClick) {
+                    Icon(
+                        imageVector = Icons.Default.AddCircle,
+                        contentDescription = "${mealType.time} 음식 추가",
+                        tint = MaterialTheme.colorScheme.primary,
+                        modifier = Modifier.size(28.dp)
+                    )
+                }
+            }
+            HorizontalDivider(
+                thickness = 1.dp,
+                color = Color.Black
+            )
+            if (addedFoods.isNotEmpty()) {
+                Column(
+                    modifier = Modifier.fillMaxWidth(),
+                    verticalArrangement = Arrangement.spacedBy(8.dp)
+                ) {
+                    addedFoods.forEach { addedFood ->
+                    }
+                }
+            } else {
+                Text(
+                    text = "아직 추가된 음식이 없습니다.",
+                    fontSize = 14.sp,
+                    color = Color.Gray,
+                    modifier = Modifier.padding(8.dp)
+                )
+            }
+        }
+    }
+}
+
+enum class MealType(val time: String) {
+    BREAKFAST("아침"),
+    LUNCH("점심"),
+    DINNER("저녁"),
+    SNACK("간식")
+}
+
+data class AddedFood(
+    val uniqueId: Long = System.currentTimeMillis(),
+    val food: Food,
+    val mealType: MealType
+)
+
+@Preview(showBackground = true)
+@Composable
+fun DietPreview() {
+    CarefullTheme {
+        Diet()
+    }
+}
