+++ app/src/main/java/com/openstudy/carefull/screen/routine/Exercise.kt
@@ -1,4 +1,4 @@
-package com.openstudy.carefull.screen
+package com.openstudy.carefull.screen.routine
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
@@ -13,19 +13,10 @@ import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.AccountBox
-import androidx.compose.material.icons.filled.AddCircle
 import androidx.compose.material3.Button
 import androidx.compose.material3.Card
 import androidx.compose.material3.CardDefaults
-import androidx.compose.material3.HorizontalDivider
-import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
 import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Scaffold
-import androidx.compose.material3.ScrollableTabRow
-import androidx.compose.material3.Tab
 import androidx.compose.material3.Text
 import androidx.compose.material3.TextButton
 import androidx.compose.runtime.Composable
@@ -37,49 +28,14 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.res.stringArrayResource
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
 import androidx.compose.ui.window.Dialog
 import androidx.compose.ui.window.DialogProperties
-import com.openstudy.carefull.R
-import com.openstudy.carefull.common.BottomNavigationBar
 import com.openstudy.carefull.ui.theme.CarefullTheme
 
-
-@Composable
-fun ExerciseAndDiet() {
-    var selectedTabIndex by remember { mutableIntStateOf(0) }
-
-    Scaffold(
-        topBar = {
-            TopNavigationBar(
-                selectedTabIndex = selectedTabIndex,
-                onTabSelected = { index, item ->
-                    selectedTabIndex = index
-                }
-            )
-        },
-        bottomBar = {
-            BottomNavigationBar(R.string.exercise_and_diet)
-        }
-    ) { innerPadding ->
-        Column(
-            modifier = Modifier
-                .fillMaxSize()
-                .padding(innerPadding),
-            horizontalAlignment = Alignment.CenterHorizontally,
-        ) {
-            when (selectedTabIndex) {
-                0 -> Exercise()
-                1 -> Diet()
-            }
-        }
-    }
-}
 
 @Composable
 fun Exercise() {
@@ -286,157 +242,11 @@ fun ExerciseCountDialog(
     }
 }
 
-@Composable
-fun TopNavigationBar(
-    selectedTabIndex: Int,                   // (1) 현재 선택된 탭의 인덱스
-    onTabSelected: (Int, String) -> Unit     // (2) 탭 선택 시 호출될 콜백
-) {
-    val navItems = stringArrayResource(id = R.array.exercise_and_diet)
-    ScrollableTabRow(
-        selectedTabIndex = selectedTabIndex,
-        modifier = Modifier.fillMaxWidth(),
-        containerColor = MaterialTheme.colorScheme.surface,
-        edgePadding = 10.dp,
-        indicator = {},
-        divider = {}
-    ) {
-        navItems.forEachIndexed { index, item ->
-            Tab(
-                selected = (selectedTabIndex == index),
-                onClick = { onTabSelected(index, item) },
-                text = {
-                    Text(
-                        text = item, style = MaterialTheme.typography.titleMedium
-                    )
-                },
-                selectedContentColor = MaterialTheme.colorScheme.primary,
-                unselectedContentColor = Color.Gray
-            )
-        }
-    }
-}
-
-@Composable
-fun Diet() {
-    var addedFoods by remember { mutableStateOf<List<AddedFood>>(emptyList()) }
-
-    Column {
-        Spacer(modifier = Modifier.height(24.dp))
-        Text(
-            text = "오늘 식사 내역", style = MaterialTheme.typography.bodyLarge,
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(horizontal = 16.dp),
-            textAlign = TextAlign.Center
-        )
-        LazyColumn(
-            modifier = Modifier
-                .fillMaxSize()
-                .padding(16.dp)
-        ) {
-            item {
-                MealType.entries.forEach { mealType ->
-                    MealSection(
-                        mealType = mealType,
-                        addedFoods = addedFoods.filter { it.mealType == mealType },
-                        onAddClick = {
-                        },
-                        onRemoveClick = { foodToRemove ->
-                        }
-                    )
-                }
-            }
-        }
-    }
-}
-
-@Composable
-fun MealSection(
-    mealType: MealType,
-    addedFoods: List<AddedFood>,
-    onAddClick: () -> Unit,
-    onRemoveClick: (AddedFood) -> Unit
-) {
-    Card(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(vertical = 12.dp),
-        elevation = CardDefaults.cardElevation(defaultElevation = 2.dp)
-    ) {
-        Column(
-            modifier = Modifier
-                .fillMaxWidth()
-        ) {
-            Row(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(8.dp),
-                verticalAlignment = Alignment.CenterVertically
-            ) {
-                Text(
-                    text = mealType.time,
-                    fontSize = 16.sp,
-                    fontWeight = FontWeight.Bold
-                )
-                Spacer(modifier = Modifier.weight(1f))
-                IconButton(onClick = onAddClick) {
-                    Icon(
-                        imageVector = Icons.Default.AccountBox,
-                        contentDescription = "${mealType.time} 음식 추가",
-                        tint = MaterialTheme.colorScheme.primary,
-                        modifier = Modifier.size(28.dp)
-                    )
-                }
-                IconButton(onClick = onAddClick) {
-                    Icon(
-                        imageVector = Icons.Default.AddCircle,
-                        contentDescription = "${mealType.time} 음식 추가",
-                        tint = MaterialTheme.colorScheme.primary,
-                        modifier = Modifier.size(28.dp)
-                    )
-                }
-            }
-            HorizontalDivider(
-                thickness = 1.dp,
-                color = Color.Black
-            )
-            if (addedFoods.isNotEmpty()) {
-                Column(
-                    modifier = Modifier.fillMaxWidth(),
-                    verticalArrangement = Arrangement.spacedBy(8.dp)
-                ) {
-                    addedFoods.forEach { addedFood ->
-                    }
-                }
-            } else {
-                Text(
-                    text = "아직 추가된 음식이 없습니다.",
-                    fontSize = 14.sp,
-                    color = Color.Gray,
-                    modifier = Modifier.padding(8.dp)
-                )
-            }
-        }
-    }
-}
-
-enum class MealType(val time: String) {
-    BREAKFAST("아침"),
-    LUNCH("점심"),
-    DINNER("저녁"),
-    SNACK("간식")
-}
-
-data class AddedFood(
-    val uniqueId: Long = System.currentTimeMillis(),
-    val food: Food,
-    val mealType: MealType
-)
 
 @Preview(showBackground = true)
 @Composable
-fun ExerciseAndDietPreview() {
+fun ExercisePreview() {
     CarefullTheme {
-        ExerciseAndDiet()
+        Exercise()
     }
 }
